backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=4.3.0",
    "beanie>=1.29.0",
    "fastapi-mail>=1.4.2",
    "fastapi[standard]>=0.115.11",
    "itsdangerous>=2.2.0",
    "pydantic-settings>=2.8.1",
    "pyjwt>=2.10.1",
]

[project.optional-dependencies]
test = [
    "faker>=37.1.0",
    "httpx>=0.28.1",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.25.3",
]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"


